Refactor Editor class to be ..
.. an abstract class. Then, MoveType, RenameFile, RenameType edits are
performed by different subclasses of the editor.
